Add option to change permissions on upload dir
Since some people serve their static data by uploads directory, unhandy
to set the permissions of the upload directory to 0700 by default.

In order to allow everyone to set the permissions themselves, an option
`UPLOADS_MODE` is introduced that will allow to define the default mode
set to the directory.

Signed-off-by: Sheogorath <sheogorath@shivering-isles.com>